# receipt-processor-challenge/src/models/receipt.go
package models

type Receipt struct {
	Retailer     string  `json:"retailer"`
	PurchaseDate string  `json:"purchaseDate"`
	PurchaseTime string  `json:"purchaseTime"`
	Items        []Items `json:"items"`
	Total        string `json:"total"`
}